backend2/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
django-cors-headers==4.7.0
firebase-admin==6.5.0
psycopg2-binary==2.9.10
google-cloud-firestore==2.20.2